requirements.txt
google-genai>=1.0.0
numpy>=1.26.0
asyncpg>=0.29.0
